gitlab-ci.yml: Avoid running CI for merge pipelines
This avoids running pipelines *after* the MR has been merged and
avoids triggering a new CI build and with it the tests.

These should already be done in the MR context so there's no need to
run them once more time.

Further more, with this change forks would need to manually trigger
builds. No change for the MR context, as these it would build automatically.

Signed-off-by: Marius Vlad <marius.vlad@collabora.com>